pyproject.toml
[tool.poetry]
name = "django-challenge-sc"
version = "0.2.0"
description = ""
authors = ["Malazay SouthernCode <alexisgiovoglanian@southerncode.us>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.1"
django-rest-framework = "^0.1.0"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.1"
django-filter = "^24.2"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
